refactor: split registry trust facade
