+++ crates/assay-registry/src/trust_next/mod.rs
@@ -1,0 +1,5 @@
+pub(in crate::trust) mod access;
+pub(in crate::trust) mod cache;
+pub(in crate::trust) mod decode;
+pub(in crate::trust) mod manifest;
+pub(in crate::trust) mod pinned;
